+++ CHANGELOG.md
@@ -36,6 +36,10 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
     *Jon Rohan*
 
+* Border arguments are now pulled in through the utilities class.
+
+    *Jon Rohan*
+
 ### Breaking changes
 
 * `bg:` system argument will no longer accept hex color strings, and deprecated color scale.